requirements.txt
torch==2.0.1
transformers==4.37.2
Pillow==8.4.0
